scheduler: add socket closing upon stopping

+++ include/work_item_scheduler.hh
@@ -36,5 +36,5 @@ private:
     bool active_ = false;
 
     /* A map to keep track of which socket fd belongs to which worker */
-    std::map<uint64_t, size_t> fd_worker_map_;
+    std::map<int, size_t> fd_worker_map_;
 };
